refactor(workspace): share reconciled eformsign types (P5.4a)
First drifted-type slice — eformsign contracts reconcile against the
backend source of truth into packages/shared/src/types/eformsign
(@babyjamjam/shared/types/eformsign):
- filtered document lists: apps assumed { total_rows, limit, skip };
  the backend emits { total_count } — shared raw envelope added and
  both service layers normalize explicitly
- mobile expected FinalizeHeadlessResponse.failedStep which the backend
  DTO never had — dropped, contracts page fallback fixed
- frontend's narrowed client summary widened to the backend shape
- SDK layout options gain zoom / object-form viewer_toolbar; stale
  header/footer dropped
- app-derived view/webhook helper models deliberately stay app-local

Also deletes dead frontend/types/api.ts (marketing-site contract;
import-grep proven unused — only doc references remain).

Gates: both type-checks clean, 630 + 245 tests green, lint counts
unchanged (0/206, 0/146), both production builds exit 0.
